[swappable] Use proxy models for device_monitoring app's models
