+++ tests/openwisp2/sample_check/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'openwisp2.sample_check.apps.SampleCheckConfig'
