+++ tests/openwisp2/sample_device_monitoring/__init__.py
@@ -1,1 +1,3 @@
-default_app_config = 'openwisp_monitoring.device.apps.DeviceMonitoringConfig'
+default_app_config = (
+    'openwisp2.sample_device_monitoring.apps.SampleDeviceMonitoringConfig'
+)
