+++ tests/openwisp2/sample_monitoring/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'openwisp2.sample_monitoring.apps.SampleMonitoringConfig'
